# apps/backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/modules": {
      "post": {
        "description": "Create a new module and link to appropriate course",
        "operationId": "ModuleController_createModule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Module created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid module payload"
          },
          "409": {
            "description": "Module code already exists for course"
          }
        },
        "summary": "Create a module",
        "tags": [
          "Modules"
        ]
      },
      "get": {
        "description": "Filter by userId(enrolled) | courseId(course owned) | universityId(modules for university over all courses). At least one filter required",
        "operationId": "getAllModules",
        "parameters": [
          {
            "name": "universityId",
            "required": false,
            "in": "query",
            "description": "Filter by university ID - returns all modules across all courses in the university",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "description": "Filter by course ID - returns all modules in the course",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "GroupID",
            "required": false,
            "in": "query",
            "description": "Filter by ModuleGrouping ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "moduleCode",
            "required": false,
            "in": "query",
            "description": "Filter by code, makes use of wildcard search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userEnrollment",
            "required": false,
            "in": "query",
            "description": "Choose to filter modules based of current user enrollments",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modules returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request - at least one filter required"
          },
          "404": {
            "description": "No modules found matching the filters"
          }
        },
        "summary": "Get all modules with filters",
        "tags": [
          "Modules"
        ]
      }
    },
...
    "/api/Courses/getAll": {
      "post": {
        "operationId": "getCourses",
        "parameters": [
          {
            "name": "Degree",
            "required": false,
            "in": "query",
            "description": "Filter by Degree",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseFilters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Courses returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all courses",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/api/events": {
      "post": {
        "operationId": "createEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid event payload"
          },
          "401": {
            "description": "No active session"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "500": {
            "description": "Event was not created"
          }
        },
        "summary": "Create an event",
        "tags": [
          "Events"
        ]
      },
      "get": {
        "description": "Apply filters - Enable stats mode with filters.stats=TRUE",
        "operationId": "getAllEvents",
        "parameters": [
          {
            "name": "moduleId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "timetableId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "all",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Stats",
            "required": false,
            "in": "query",
            "description": "Enable stats = TRUE",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponseDtoV2"
                }
              }
            }
          },
          "401": {
            "description": "No active session"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "summary": "Get all events",
        "tags": [
          "Events"
        ]
      }
    },
...
    "/api/api-service/courses": {
      "get": {
        "description": "Fetches courses for the authenticated user's university.",
        "operationId": "ApiServiceController_getCourses",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Courses per page",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page index: zero index",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The authenticated user is not associated with a university."
          },
          "404": {
            "description": "The university or its API adapter could not be found."
          }
        },
        "summary": "Fetch courses",
        "tags": [
          "ApiService"
        ]
      }
    },
    "/api/api-service/course": {
      "get": {
        "description": "Returns the specified course together with its module and events.",
        "operationId": "ApiServiceController_getCourse",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "query",
            "description": "UUID of the course to fetch.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseSingleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The authenticated user is not associated with a university."
          },
          "404": {
            "description": "The university or its API adapter could not be found."
          }
        },
        "summary": "Get a specific course",
        "tags": [
          "ApiService"
        ]
      }
    },
    "/api/api-service/modules": {
      "get": {
        "description": "Fetches modules for a course at the authenticated user's university.",
        "operationId": "ApiServiceController_getModules",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "query",
            "description": "UUID of the course to fetch modules for.",
            "schema": {
              "example": "00000000-0000-0000-0000-000000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modules fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The university or course ID is invalid."
          },
          "404": {
            "description": "The university, course, or API adapter could not be found."
          }
        },
        "summary": "Fetch modules",
        "tags": [
          "ApiService"
        ]
      }
    },
    "/api/api-service/events": {
      "get": {
        "description": "Fetches events for a module at the authenticated user's university.",
        "operationId": "ApiServiceController_getEvents",
        "parameters": [
          {
            "name": "moduleId",
            "required": true,
            "in": "query",
            "description": "UUID of the module to fetch events for.",
            "schema": {
              "example": "00000000-0000-0000-0000-000000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The university or module ID is invalid."
          },
          "404": {
            "description": "The university, module, or API adapter could not be found."
          }
        },
        "summary": "Fetch events",
        "tags": [
          "ApiService"
        ]
      }
    }
  },
  "info": {
    "title": "UMTAS API",
    "description": "University Management Timetable & Scheduling System API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://capstone-vigil.dns.net.za",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "umtas-session",
        "description": "Session cookie (set automatically by BetterAuth)"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AcademicSemester": {
        "type": "string",
        "enum": [
          "SEMESTER_1",
          "SEMESTER_2",
          "YEAR"
        ],
        "description": "Academic period in which this module is taught. Null or omitted means the whole academic year."
      },
      "StylingDto": {
        "type": "object",
        "properties": {
          "colour": {
            "type": "string",
            "example": "#3B82F6"
          }
        },
        "required": [
          "colour"
        ]
      },
      "CreateCourseModuleDto": {
        "type": "object",
        "properties": {
          "Core": {
            "type": "boolean",
            "example": true,
            "description": "Identifies wether or not module is a core module of the course"
          },
          "SemesterOfStudy": {
            "type": "string",
            "nullable": true,
            "example": "Semester 1",
            "description": "Identifies which part of the year the module takes part in for the course"
          },
          "YearOfStudy": {
            "nullable": true,
            "type": "number",
            "example": 1,
            "description": "Which year does the module belong to for the course"
          }
        }
      },
      "CreateModuleDto": {
        "type": "object",
        "properties": {
          "moduleCode": {
            "type": "string",
            "minLength": 2,
            "maxLength": 15,
            "example": "COS332",
            "description": "Module code used by the university"
          },
          "moduleName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Computer Networks",
            "description": "Name of the module"
          },
          "moduleDescription": {
            "type": "string",
            "nullable": true,
            "example": "Introduction to computer networking concepts",
            "description": "Short module description"
          },
          "semester": {
            "nullable": true,
            "example": "SEMESTER_1",
            "description": "Academic period in which this module is taught. Null or omitted means the whole academic year.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AcademicSemester"
              }
            ]
          },
          "styling": {
            "nullable": true,
            "description": "Styling to be used for a Module",
            "example": {
              "colour": "FFFFF"
            },
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StylingDto"
              }
            ]
          },
          "validated": {
            "type": "boolean",
            "example": true,
            "description": "Whether the module has been approved by a university admin"
          },
          "ExternalID": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "example": "12345",
            "description": "Refer to module on external API"
          },
          "ModuleGroupingID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "ModuleGroupingID to identify group the module belongs to"
          },
          "CourseID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "CourseID module belongs to"
          },
          "CourseModuleInfo": {
            "description": "Course Module metadata to be used when module belongs to course",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateCourseModuleDto"
              }
            ]
          }
        },
        "required": [
          "moduleCode",
          "moduleName"
        ]
      },
      "CourseModuleDto": {
        "type": "object",
        "properties": {
          "CourseModuleID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "ID to identify Course Module entry"
          },
          "GroupModuleID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "ID to identify GroupModule entry metadata is for"
          },
          "CourseID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "ID to identify course owning this CourseModule"
          },
          "Core": {
            "type": "boolean",
            "example": true,
            "description": "Identifies wether or not module is a core module of the course"
          },
          "SemesterOfStudy": {
            "type": "string",
            "nullable": true,
            "example": "Semester 1",
            "description": "Identifies which part of the year the module takes part in for the course"
          },
          "YearOfStudy": {
            "nullable": true,
            "type": "number",
            "example": 1,
            "description": "Which year does the module belong to for the course"
          }
        },
        "required": [
          "CourseModuleID",
          "GroupModuleID",
          "CourseID"
        ]
      },
...
      "VenueDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "venueName": {
            "type": "string"
          }
        },
        "required": [
          "venueId",
          "venueName"
        ]
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "activityType": {
            "type": "string",
            "enum": [
              "lecture",
              "tutorial",
              "prac",
              "test",
              "exam"
            ]
          },
          "eventCriteria": {
            "$ref": "#/components/schemas/EventCriteriaDto"
          },
          "isRecurring": {
            "type": "boolean"
          },
          "validated": {
            "type": "boolean"
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueDto"
            }
          }
        },
        "required": [
          "eventId",
          "eventCriteria"
        ]
      },
      "ModuleSingleResponseDto": {
        "type": "object",
        "properties": {
          "moduleID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a module"
          },
          "moduleCode": {
            "type": "string",
            "minLength": 2,
            "maxLength": 15,
            "example": "COS332",
            "description": "Module code used by the university"
          },
          "moduleName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Computer Networks",
            "description": "Name of the module"
          },
          "moduleDescription": {
            "type": "string",
            "nullable": true,
            "example": "Introduction to computer networking concepts",
            "description": "Short module description"
          },
          "semester": {
            "nullable": true,
            "example": "SEMESTER_1",
            "description": "Academic period in which this module is taught. Null or omitted means the whole academic year.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AcademicSemester"
              }
            ]
          },
          "styling": {
            "nullable": true,
            "description": "Styling to be used for a Module",
            "example": {
              "colour": "FFFFF"
            },
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StylingDto"
              }
            ]
          },
          "CourseModuleInfo": {
            "nullable": true,
            "description": "Metadata attached to module when owned by course",
            "example": {
              "CourseModuleID": "00000000-0000-0000-0000-000000000000",
              "GroupModuleID": "00000000-0000-0000-0000-000000000000",
              "CourseID": "00000000-0000-0000-0000-000000000000",
              "core": true,
              "SemesterOfStudy": "Semester 1",
              "YearOfStudy": 1
            },
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseModuleDto"
              }
            ]
          },
          "validated": {
            "type": "boolean",
            "example": true,
            "description": "Whether the module has been approved by a university admin"
          },
          "Events": {
            "nullable": true,
            "description": "List of events for the module",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "ExternalID": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "example": "12345",
            "description": "Refer to module on external API"
          },
          "Enrolled": {
            "type": "boolean",
            "example": false,
            "description": "Is a user enrolled to this module"
          },
          "ModuleGroupingID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a module group"
          }
        },
        "required": [
          "moduleID",
          "moduleCode",
          "moduleName"
        ]
      },
      "ModuleListResponseDto": {
        "type": "object",
        "properties": {
          "modules": {
            "description": "List of modules",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleSingleResponseDto"
            }
          },
          "message": {
            "type": "string",
            "description": "Short message indicating  success of response."
          }
        },
        "required": [
          "modules"
        ]
      },
...
      "CourseSingleResponseDto": {
        "type": "object",
        "properties": {
          "CourseID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a course"
          },
          "UniversityID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a university"
          },
          "GroupID": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a group of modules"
          },
          "CourseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Computer Science",
            "description": "Name of the course"
          },
          "Degree": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 30,
            "example": "Bachelor of Science",
            "description": "Degree that course belongs to"
          },
          "Modules": {
            "description": "Modules for the course.",
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModulesDto"
            }
          },
          "ExternalID": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 255,
            "example": "12345",
            "description": "Refer to course on external API"
          }
        },
        "required": [
          "CourseID",
          "UniversityID",
          "CourseName"
        ]
      },
      "CourseFilters": {
        "type": "object",
        "properties": {
          "UniversityID": {
            "type": "string",
            "format": "uuid",
            "example": "00000000-0000-0000-0000-000000000000",
            "description": "Unique identifier for a university"
          },
          "CourseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Computer Science",
            "description": "Name of the course"
          },
          "Degree": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 30,
            "example": "Bachelor of Science",
            "description": "Degree that course belongs to"
          }
        }
      },
...
      "CourseListResponseDto": {
        "type": "object",
        "properties": {
          "courses": {
            "description": "List of courses",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "courses"
        ]
      },
...
      "EventSingleResponseDto": {
        "type": "object",
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "event"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "activityCode": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 10
          },
          "eventCriteria": {
            "$ref": "#/components/schemas/EventCriteriaDto"
          },
          "isRecurring": {
            "type": "boolean"
          },
          "validated": {
            "type": "boolean"
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueDto"
            }
          },
          "activityType": {
            "type": "string",
            "enum": [
              "lecture",
              "tutorial",
              "prac",
              "test",
              "exam"
            ],
            "description": "Required when eventCriteria.moduleId is provided."
          }
        },
        "required": [
          "eventCriteria"
        ]
      },
      "EventListResponseDtoV2": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "count": {
            "type": "number",
            "example": 68,
            "description": "Count of Modules"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "events"
        ]
      },
...
      "EventListResponseDto": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "events"
        ]
      }
    }
  }
}